casi listo

+++ BaseT0/src/lrsh/process.h
@@ -3,11 +3,11 @@
 #include <sys/types.h>
 
 typedef struct {
-    int pid;  // or pid_t if it's available after including the correct header
+    int pid;  
     char* name;
     int exit_code;
-    double start_time;
-    double end_time;
+    int start_time;
+    int end_time;
 } Process;
 
 void update_process_exit_code(pid_t pid, int exit_code, Process* processes ,int* count);
